Refactor folder structure
Fixes #27

Refactor the codebase into a modular structure.

* Create `modules/invoice` directory and move all invoice-related files from `src/components/ui/invoice` to `modules/invoice`.
  * Move `BillingInfo.tsx`, `EntriesTable.tsx`, `EntryRow.tsx`, `GenerateInvoice.tsx`, `InvoiceFooter.tsx`, `InvoiceHeader.tsx`, `TaxDetail.tsx`, and `TaxDetailsTable.tsx` to appropriate subdirectories within `modules/invoice`.
* Create `modules/auth` directory for future authentication-related files.
* Update imports in `src/app/generate/page.tsx` to reflect the new modular structure.
* Delete the original files from `src/components/ui/invoice`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/keizerworks/invoicen/issues/27?shareId=XXXX-XXXX-XXXX-XXXX).
